+++ core/httpd-platform.h
@@ -1,7 +1,7 @@
 #ifndef HTTPD_PLATFORM_H
 #define HTTPD_PLATFORM_H
 
-#include "platform.h"
+#include "libesphttpd/platform.h"
 
 int httpdPlatSendData(ConnTypePtr conn, char *buff, int len);
 void httpdPlatDisconnect(ConnTypePtr conn);
