+++ espfs/mkespfsimage/main.c
@@ -9,7 +9,7 @@
 #ifdef __MINGW32__
 #include <io.h>
 #endif
-#include "espfs.h"
+#include "libesphttpd/espfs.h"
 #include "espfsformat.h"
 
 //Heatshrink
