+++ include/libesphttpd/auth.h
