+++ include/libesphttpd/cgiwifi.h
