+++ util/cgiwebsocket.c
@@ -12,11 +12,11 @@ Websocket support for esphttpd. Inspired by https://github.com/dangrie158/ESP-82
  */
 
 
-#include <esp8266.h>
-#include "httpd.h"
-#include "sha1.h"
+#include <libesphttpd/esp8266.h>
+#include "libesphttpd/httpd.h"
+#include "libesphttpd/sha1.h"
 #include "base64.h"
-#include "cgiwebsocket.h"
+#include "libesphttpd/cgiwebsocket.h"
 
 #define WS_KEY_IDENTIFIER "Sec-WebSocket-Key: "
 #define WS_GUID "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"
